+++ problem/set_cover.go
@@ -3,6 +3,7 @@ package problem
 import (
 	"github.com/roidaradal/fn/dict"
 	"github.com/roidaradal/fn/list"
+	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
 )
@@ -13,16 +14,18 @@ func NewSetCover(variant string, n int) *discrete.Problem {
 	switch variant {
 	case "basic":
 		return setCover(name)
+	case "weighted":
+		return weightedSetCover(name)
 	default:
 		return nil
 	}
 }
 
-// Set Cover
-func setCover(name string) *discrete.Problem {
+// Create new Set Cover problem
+func newSetCoverProblem(name string) (*discrete.Problem, *data.Subsets) {
 	p, cfg := newSubsetsProblem(name)
 	if p == nil || cfg == nil {
-		return nil
+		return nil, nil
 	}
 
 	p.Goal = discrete.Minimize
@@ -37,5 +40,25 @@ func setCover(name string) *discrete.Problem {
 		}
 		return list.AllTrue(dict.Values(covered))
 	})
+	return p, cfg
+}
+
+// Set Cover
+func setCover(name string) *discrete.Problem {
+	p, _ := newSetCoverProblem(name)
+	return p
+}
+
+// Weighted Set Cover
+func weightedSetCover(name string) *discrete.Problem {
+	p, cfg := newSetCoverProblem(name)
+	if p == nil || cfg == nil {
+		return nil
+	}
+	if len(cfg.Weight) != len(cfg.Names) {
+		return nil
+	}
+
+	p.ObjectiveFn = fn.ScoreSumWeightedSubset(cfg.Names, cfg.Weight)
 	return p
 }